Moves everything in app container to a listContainer and skeletons out where Listings List will go
