+++ src/components/ListingsContainer/index.jsx
@@ -1,0 +1,56 @@
+import React, { Component } from 'react';
+import client from '../../services/client';
+import styles from './styles.scss';
+
+import Header from '../common/Header';
+import NewListingForm from '../NewListingForm';
+
+class ListingsContainer extends Component {
+  constructor() {
+    super();
+    this.state = {
+      listings: [],
+      loading: true,
+      showServerError: false
+    };
+  }
+
+  componentDidMount() {
+    return client.getListings().then(listings => {
+      return this.setState({ listings: listings, loading: false, showServerError: false });
+    }).catch(() => {
+      return this.setState({ showServerError: true });
+    });
+  }
+
+  handleSubmit(title, url) {
+    let listings = this.state.listings.slice();
+
+    return client.createListing(title, url).then(listing => {
+      listings.push(listing);
+
+      return this.setState({
+        listings: listings,
+        showServerError: false
+      });
+    });
+  }
+
+  render() {
+    return(
+      <div className={styles.container}>
+      <Header title="Listings" />
+        <main>
+          <NewListingForm
+            className={styles.newListingForm}
+            onSubmit={(title, url) => this.handleSubmit(title, url)}
+          />
+          <div>This is where the Listings List will Go</div>
+        </main>
+      </div>
+    );
+  }
+
+}
+
+export default ListingsContainer;
