+++ src/components/ListingsContainer/styles.scss
